アクセストークンページを追加し、JWTトークン生成を実装。authOptionsを設定し、必要な依存関係をpackage.jsonに追加。
