+++ my-next-app/app/config/authOptions.ts
@@ -1,0 +1,26 @@
+import GoogleProvider from "next-auth/providers/google";
+import { AuthOptions } from "next-auth";
+
+export const authOptions: AuthOptions = {
+  providers: [
+    GoogleProvider({
+      clientId: process.env.GOOGLE_CLIENT_ID!,
+      clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
+    }),
+  ],
+  secret: process.env.NEXTAUTH_SECRET,
+  callbacks: {
+    async jwt({ token, user, account }) {
+      if (user && account) {
+        token.access_token = account.access_token;
+      }
+
+      return token;
+    },
+
+    async session({ session, token }) {
+      session.token = token.access_token || '';
+      return session;
+    }
+  }
+}
